+++ src/Smartstore.Core/Catalog/Products/Import/ProductImporter.cs
@@ -413,7 +413,7 @@ namespace Smartstore.Core.DataExchange.Import
                 row.SetProperty(context.Result, (x) => x.AvailableForPreOrder);
                 row.SetProperty(context.Result, (x) => x.CallForPrice);
                 row.SetProperty(context.Result, (x) => x.Price);
-                row.SetProperty(context.Result, (x) => x.OldPrice);
+                row.SetProperty(context.Result, (x) => x.ComparePrice);
                 row.SetProperty(context.Result, (x) => x.ProductCost);
                 row.SetProperty(context.Result, (x) => x.SpecialPrice);
                 row.SetProperty(context.Result, (x) => x.SpecialPriceStartDateTimeUtc);